Move command line parser to UI #33
- refactoring the code to prepare further modifications.
